drug-discovery/individual/hangler/practical_work/archiv/downstream_tasks/requirements.txt
rdkit-pypi
boto3
pytorch-ignite
gpytorch
git+https://github.com/y0ast/DUE.git
git+https://github.com/ml-jku/clamp
numpy
